Do the parallelization in a separate interpeter
This adds the `Join` construct to the underlying Fetch structure.
The `Join`s are paralellized in a new interpreter `parallelJoinPhase`.
The new parallel optimization fiddles with the `Free` structure to look
for independent queries and to simplify queries if the results are
cached. Everything in a `fetch.map(...)` or `fetch.flatMap(...)` should
be only executed once (currently this isn't yet the case for pure
fetches created with `Fetch.pure`).