Add twitter future support (#128)
* initial cut at twitter support

this is jvm only, so I'm not sure exactly how to make the work here
but the core of this works. Twitter future and Catbirds Rerunnable are
now supported

* make the thing build

and execute the tests to verify they're passing

* add an evalToRerunnable

this matches the monix side

* spell rerunnable the conventional way

with two n's in ht middle

* add implicit FuturePool

with the default value of the interruptableUnboundedPool and then use
the fact that the pool creates interruptable threads to wire up the timeouts.

* simplify dependencies

rely on `catbird` to provide the twitter `util-core` dependency.

* remove 2.10 cross version

Twitter util no longer supports 2.10 so this won't build with it in. also remove the now unneeded version check for including paradise.

* remove 2.10

oops, missed travis

* extend test coverage

test evalToRunnable

* test sync

for both Rerunnable and Future ... the code is there and it should
work so prove it.

* add applicative usage tests

for both twitter future and catbirds rerunnable

* update twitter future Ap handler

to use the catbird provided ap

* use evalToRerunnable

in the rerunnable sync case

* add test to excercise twfuture fetch timeouts

adapted from the scala future version

* also test Rerunnable side

when testing timeouts