Add `F[_]` type parameter to `DataSource`
In the old version of Fetch, one needed to define the data source as an
implicit object. This way, we ensured that each DataSource had only one instance using reference equality.

A data source used to look like the following:

```scala
implicit object Users extends DataSource[UserId, User]{
  // ...
}
```

Now that we've introduced an `F[_]` type parameter in `DataSource` we no
longer can define it as an `object`. Thus, we need a way to identify
requests to a data source to be able to perform optimizations.

Now there's a new trait called `Data`, that identifies requests for a
particular piece of data, and is used to identify and group requests for
the same data.

A `Data` definition looks like the following:

```scala
object Users extends Data[Int, User] {
  def name = "Users"
}
```

Then, when defining the data source for users, we have to specify which
`Data` the `DataSource` is fetching:

```scala
def usersSource[F[_]]: DataSource[F, Int, User] = new DataSource[F, Int,
User] {
  def data = Users

  // ...
}
```

Now that we have `Data` and `DataSource` we can create a `Fetch` por a `User`:

```scala
def fetchUser(id: Int): Fetch[F, User] =
  Fetch(id, usersSource)
```

See the `examples` directory and the tests for plenty of examples.