Remove Unfetch wrapper class.
This commit removes the Unfetch case class, a wrapper for a
F[FetchResult], and reduces the `Fetch` trait into a mere type alias.

The `Unfetch` class, as defined, is a mere artifice used to provide
an instance of the `Monad` type-class for a `F[FetchResult]`.

Instead of creating a wrapper class, and a Monad instance for that
wrapper class, in this PR we simply create a more specific monad
instance for an `F[FetchResult]`, one which handles both the `F` and
the inner `FetchResult` ADT structure.